+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Breaking Changes
 - Removed `LinearNormalActionNoise`
+- Evaluation is now deterministic by default, except for Atari games
 
 ### New Features
 - Added option to choose which `VecEnv` class to use for multiprocessing
@@ -19,6 +20,7 @@
 - Use f-strings when possible
 - Changed `PPO` atari hyperparameters (removed vf clipping)
 - Changed `A2C` atari hyperparameters (eps value of the optimizer)
+- Updated benchmark script
 
 ## Pre-Release 0.10.0 (2020-10-28)